_Archive/requirements.txt
crewai==0.16.3
langchain==0.1.6
langchain-community==0.0.20
langsmith>=0.0.83,<0.1
crewai
langchain-openai
pydantic
python-dotenv
streamlit
openai